moved frontend login part into auth
